refactor: clarify session connection state naming
